kaggle-brazilian-ecommerce-olist/pyproject.toml
[tool.poetry]
name = "kaggle-brazilian-ecommerce-olist"
version = "0.1.0"
description = ""
authors = ["Pricilla Patricia de Aragao <pricillapatricia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
